src/requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
attrs==22.1.0
black==22.10.0
click==8.1.3
exceptiongroup==1.0.4
Faker==15.3.4
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
